requirements.txt
python-telegram-bot[all]==21.8
aiohttp==3.10.10
python-dotenv==1.0.1
pymongo==4.6.1
